requirements.txt
fastapi
uvicorn
pandas
python-multipart
sentence-transformers
python-docx
openpyxl
python-pptx
pdfplumber
python-dotenv
faiss-cpu
numpy
ollama
cryptography
groq
keyring